+++ Final Project-Routing with Graph Algorithms/Point.cpp
@@ -3,7 +3,7 @@
 #include "Headers/Point.h"
 
 double Point::CalculateDistance(Point target) {
-	int dx = this->x - target.x;
-	int dy = this->y - target.y;
+	double dx = this->x - target.x;
+	double dy = this->y - target.y;
 	return sqrt(dx * dx + dy * dy);
 }
